--- a/skyplane/compute/gcp/gcp_server.py
+++ b/skyplane/compute/gcp/gcp_server.py
@@ -51,6 +51,9 @@ class GCPServer(Server):
         """Get public IP for instance with GCP client"""
         return self.get_instance_property("networkInterfaces")[0]["accessConfigs"][0].get("natIP")
 
+    def private_ip(self):
+        return self.get_instance_property("networkInterfaces")[0]["networkIP"]
+        
     def instance_class(self):
         machine_type = self.get_instance_property("machineType")
         return machine_type.split("/")[-1] if machine_type else None