PR for #585. 
* add private_ip() for AWS and GCP
* use private ips for GCP to GCP data transfer
* Also add fix for #603 

Private ips for GCP throughput profiling is in #605