@parasj 2022-10-13
Can you profile a typical GCP to GCP transfer before+after this change?