@lynnliu030 2022-10-13
@parasj Here's an example of transferring 9GB of data from gcp to gcp. 

Before (using public ips): transfer runtime = 17.18s, throughput = 4.23 Gbps 
<img width="1223" alt="image" src="https://user-images.githubusercontent.com/39693493/195473510-4dae3402-45c0-4dbc-845d-23c1e748e54d.png">

After (using private ips): transfer runtime = 10.29s, throughput = 7.06 Gbps 
<img width="1188" alt="image" src="https://user-images.githubusercontent.com/39693493/195473750-0bf0e103-0b44-4044-a889-dc238c04f248.png">

There's around 1.7x throughput improvement and 40% runtime reduction